Move readouts from columns to end explanation
